Refactors image sending to use shared logic
Moves the image sending logic into a shared function
to be used by both the remote server and the collab host.

This change allows both remote and collab sessions to use
the same image chunking and sending logic since the client
handlers are the same anyway and expect the same chunking
behavior.